Move logic analyzer to separate module
